- Fixed wrong type for m_steps and m_done. - Registration with the job tracker is now done earlier.
svn path=/trunk/playground/network/smb4k/; revision=1224135

+++ core/smb4kprint_p.h
@@ -140,8 +140,8 @@ class Smb4KPrintJob : public KJob
     Smb4KShare *m_share;
     QWidget *m_parent_widget;
     QString m_temp_dir;
-    int m_steps;
-    int m_done;
+    qulonglong m_steps;
+    qulonglong m_done;
 };
